requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.3
Flask==2.3.3
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.30.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
-e git+https://github.com/ultralytics/ultralytics@52afadcb9772fb17933f02dc646b00074c184fba#egg=ultralytics
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
zipp==3.16.2
